src/py_load_chembl/adapters/postgres.py
import psycopg2
from typing import Any, Dict, List
from pathlib import Path

from py_load_chembl.adapters.base import DatabaseAdapter

class PostgresAdapter(DatabaseAdapter):
    """Database adapter for PostgreSQL."""

    def __init__(self, connection_string: str):
        self.connection_string = connection_string
        self.conn = None

    def connect(self) -> Any:
        """Establishes a connection to the target database."""
        if self.conn is None or self.conn.closed:
            try:
                self.conn = psycopg2.connect(self.connection_string)
            except psycopg2.OperationalError as e:
                raise ConnectionError(f"Failed to connect to PostgreSQL: {e}") from e
        return self.conn

    def execute_ddl(self, ddl_script: str) -> None:
        """Executes a DDL script (e.g., schema creation, index management)."""
        pass

    def bulk_load_table(self, table_name: str, data_source: Path | str, schema: str = "public", options: Dict[str, Any] = {}) -> None:
        """
        Performs high-performance native bulk loading into a specific table.
        """
        pass

    def execute_merge(self, source_table: str, target_table: str, primary_keys: List[str], columns: List[str]) -> None:
        """
        Executes an efficient MERGE/UPSERT operation from source (staging) to target (production).
        """
        pass

    def optimize_pre_load(self, schema: str) -> None:
        """Disables constraints and indexes before a full load."""
        pass

    def optimize_post_load(self, schema: str) -> None:
        """Re-enables constraints, rebuilds indexes, and runs statistics gathering (e.g., ANALYZE)."""
        pass
